# Optimize SSE/SSE2 Set[All]Vector128