… src/jit/hwintrinsiclistxarch.h
… src/jit/hwintrinsicxarch.cpp
… src/jit/lsraxarch.cpp
… tests/src/JIT/HardwareIntrinsics/X86/Sse2/SetVector128.cs
… tests/src/JIT/HardwareIntrinsics/X86/Sse2/SetVector128_r.csproj
… tests/src/JIT/HardwareIntrinsics/X86/Sse2/SetVector128_ro.csproj