This PR moves the implementation of SSE/SSE2 `SetVector128` and `SetAllVector128` from managed code to JIT inside.
1. leveraging new instructions to get smaller code size on new CPUs. This optimization cannot be achieved by managed code because the large function body will block inlining.
2. eliminate the unnecessary "zero-upper-vector" code of `SetScalarVector128`.

Most of the `SetAllVector128` intrinsics get obvious performance improvement.

| `SetAllVector128<byte>` | AVX2 | SSE4.1 | SSE2 |
|------------|---------|----------|--------------|
| Baseline  | 4.9s | 4.93s | 4.93s |
| Optimized | 2.61s | 2.44s | 4.86s |
| Performance gain | 46.7% | 50.5% | 1.4% |

| `SetAllVector128<short>` | AVX2 | SSE4.1 | SSE2 |
|----------------------------|-------|---------|------|
| Baseline  | 3.61s | 3.61s | 3.61s |
| Optimized | 2.41s | 3.61s | 3.6s |
| Performance gain | 33% | 0% | 0.2% |

| `SetAllVector128<int>` | AVX2 | SSE4.1 | SSE2 |
|----------------------------|-------|---------|------|
| Baseline  | 2.47s | 2.41s | 2.41s |
| Optimized | 2.41s | 2.42s | 2.41s|
| Performance gain | 2.4% | 0.04% | 0% |

| `SetAllVector128<long>` | AVX2 | SSE4.1 | SSE2 |
|----------------------------|-------|---------|------|
| Baseline  | 2.44s | 2.42s | 2.42s |
| Optimized | 2.41s | 2.41s | 2.4s |
| Performance gain | 1.2% | 0.4% | 0.8% |

| `SetAllVector128<float>` | AVX2 | SSE4.1 | SSE2 |
|----------------------------|-------|---------|------|
| Baseline  | 2.42s | 2.42s | 2.42s |
| Optimized | 1.22s | 1.5s | 1.5s |
| Performance gain | 49.6% | 38% | 38% |

| `SetAllVector128<double>` | AVX2 | SSE4.1 | SSE2 |
|----------------------------|-------|---------|------|
| Baseline  | 2.43s | 2.46s | 2.46s |
| Optimized | 1.2s | 1.22s | 1.52s |
| Performance gain | 50.6% | 50.4% | 38.2% |


`SetVector128` gets significant improvement on `float` and `double`. The new implementation of `SetVector128<byte/sbyte>` makes performance regression (~5%) in the micro-benchmark, but it avoids the `stackalloc` and related stack zeroing that is very expensive sometimes.

| `SetVector128<byte>` | AVX2 | SSE4.1 | SSE2 |
|----------------------------|-------|---------|------|
| Baseline  | 35.58s | 35.26s | 35.26s |
| Optimized | 37.47 | 37.52 | 37.4s |
| Performance gain | -5.3% |  -6.4% | -6.1% |

| `SetVector128<short>` | AVX2 | SSE4.1 | SSE2 |
|----------------------------|-------|---------|------|
| Baseline  | 18.2s  | 18.3s | 18.3s  |
| Optimized | 18.18s | 18.19s | 18.18s |
| Performance gain | 0.1% | 0.6% | 0.7% | 

| `SetVector128<int>` | AVX2 | SSE4.1 | SSE2 |
|----------------------------|-------|---------|------|
| Baseline  | 8.46s | 8.56s | 8.56s |
| Optimized | 8.51s | 8.48s | 8.46s |
| Performance gain | -0.6% | 0.9% | 1% |

| `SetVector128<long>` | AVX2 | SSE4.1 | SSE2 |
|----------------------------|-------|---------|------|
| Baseline  | 3.67s | 3.64s | 3.64s |
| Optimized | 3.62s | 3.62s | 3.60s |
| Performance gain | 1.4% | 0.5% | 1% |

| `SetVector128<float>` | AVX2 | SSE4.1 | SSE2 |
|----------------------------|-------|---------|------|
| Baseline  | 8.52s | 8.54s | 8.58s |
| Optimized | 3.67s | 3.66s | 3.62s |
| Performance gain | 56.9% | 57.1% | 58% |

| `SetVector128<double>` | AVX2 | SSE4.1 | SSE2 |
|----------------------------|-------|---------|------|
| Baseline  | 3.92s | 3.66s | 3.66s |
| Optimized | 2.44s | 2.42s | 2.43s |
| Performance gain | 37.8% | 33.9% | 33.6% |

Closes #18436