@tannergooding 2018-06-22
I'm happy to see the perf gains, but as I called out on https://github.com/dotnet/coreclr/issues/18436, I don't think moving the implementations back into the JIT is the correct answer here.

This is a perfect example of "inefficient codegen" and it should be fixable by ensuring:
* We have the correct APIs exposed
* We are emitting good assembly for the exposed APIs

> This optimization cannot be achieved by managed code because the large function body will block inlining

Marking the method with `AggressiveInlining` should resolve this.

> eliminate the unnecessary "zero-upper-vector" code of SetScalarVector128.

We should be able to make the JIT smart enough to skip this in various circumstances. This is a good opportunity for a peephole optimization, but it could also just start with a simple heuristic of `SetScalarVector128` being directly consumed by another HWIntrinsic API that only reads the lower bits (for example `Sse.MoveScalar` takes `upper` and `value`. If `SetScalarVector128` is passed to `value`, we don't need to zero the upper bits).