@tannergooding 2018-06-22
To clarify on the pushback:

End-users will only have access to the exposed APIs (and cannot codify custom support in the JIT), so if we can't get the managed implementations of the various helper methods (like `SetAllVector128`) good enough (using the same APIs as them), then we will certainly be getting bugs from the same end-users later saying the codegen or perf is bad.

The helper methods (generally speaking) are all examples of code that end users might be writing to support their algorithms (building off the base APIs that actually map to instructions).