@mikedn 2018-06-22
>  I don't think moving the implementations back into the JIT is the correct answer here. This is a perfect example of "inefficient codegen" and it should be fixable by ensuring:

This has been discussed before. There is no way to properly implement these in managed code due to the all constant argument special case. And it doesn't make sense to try to implement these in both managed code in the JIT and switch between implementation depending on arguments.