@tannergooding 2018-06-22
> This has been discussed before. There is no way to properly implement these in managed code due to the all constant argument special case

Do you mean the case where all four inputs to `SetVector128` are constant?

If this was the issue attempting to be resolved, then it should have been called out in the bug or top post (and this PR would either be the first step towards that, or would contain code that was relevant to that problem).

As is, this PR looks to be attempting to workaround issues that should instead be tracked and fixed.

> And it doesn't make sense to try to implement these in both managed code in the JIT and switch between implementation depending on arguments.

Agreed.