@fiigii 2018-06-22
> Marking the method with AggressiveInlining should resolve this.

Unfortunately, conditional codegen (`if (Avx.IsSupported) { ... } else { ... }`) will fail the inlining even though with `AggressiveInlining`.

> We should be able to make the JIT smart enough to skip this in various circumstances. This is a good opportunity for a peephole optimization,

Indeed, we should optimize `SetScalarVector128` to eliminate the unnecessary zeroing code. However, it is hard, and I do not think peephole optimization is the proper approach. The general optimization needs global analysis to make sure all the following use sites not accessing the upper bits of the vectors, and it won't be always effective.

On the other hand, the performance of `SetVector128/256` and `SetAllVector128/256` is very important. Especially `SetAllVector128/256`, x86 does not have the instructions to directly implement `Multiply(vector, scalar)`, so we have to broadcast the `scalar` value into a vector. This is a very common operation so that we should provide the best performance as soon as possible.