@AndyAyersMS 2018-06-22
@fiigii do you know why inlining is blocked in the example you mention above?