@fiigii 2018-06-22
@AndyAyersMS IIRC, that was caused by the IL size of conditional codegen (`if (Avx.IsSupported) { ... } else { ... }`), but I made the experiment at a long time ago (when the managed implementation was submitting), let me check it again.