@4creators 2018-06-25
IMO @tannergooding position of prioritizing fixing general codegen for HW intrinsics rather than fixing special case HW intrnsic usage is correct.

We should avoid leaving holes in JIT which will result in sub optimal codegen in any scenario which would involve similar helper HW intrinsic coding patterns.