@CarolEidt 2018-06-25
I agree with both @tannergooding and @4creators that we should fix the codegen issues rather than moving the implementation to the JIT. First we need to identify the specific issues that are impacting performance, and then we can analyze and prioritize them. Barring some significant analysis and discussion I cannot support moving the implementation to the JIT.

I suspect that @tannergooding is right that these issues (or most of them) are ones that will impact users of the intrinsics, and so they are better addressed by fixing codegen.