@fiigii 2018-06-26
@CarolEidt Thank you for the comments and sorry for the delay to reply. I agree that we should finally fix the codegen issues that detected from `Set[All]Vector128/256`. 
1.  Moving scalar floating-point values into vectors (`SetScalarVector128`) has unnecessary overhead from "zeroing upper bits".
2. The helper intrinsics can benefit from using new ISA instructions, for example
```csharp
        [MethodImpl(MethodImplOptions.AggressiveInlining)]
        public static Vector128<byte> SetAllVector128(byte value)
        {
            Vector128<byte> vector1 = Sse.StaticCast<uint, byte>(ConvertScalarToVector128UInt32(value));

            if (Avx2.IsSupported)
            {
                return Avx2.BroadcastScalarToVector128<byte>(vector1);
            }
            else if (Ssse3.IsSupported)
            {
                return Ssse3.Shuffle(vector1, SetZeroVector128<byte>());
            }
            else
            {
                Vector128<ushort> tmpVector1 = Sse.StaticCast<byte, ushort>(UnpackLow(vector1, vector1));
                Vector128<uint> tmpVector2 = Sse.StaticCast<ushort, uint>(UnpackLow(tmpVector1, tmpVector1));
                return Sse.StaticCast<uint, byte>(Shuffle(tmpVector2, 0));
            }
            
        }
```

For (1), we need additional analysis (in lowering?) to improve the codegen by directly using `float/double` as `Vector128<float/double>` if there is no upper-bits access on the vector. For (2), it is easy to fix in either managed code or JIT inside. BTW, I investigated the conditional code path again and detected that the managed-implemented helpers can be inlined with `AggressiveInlining` (I might use the debug S.P.CoreLib before, my apologies) cc @AndyAyersMS.

However, as @mikedn pointed, the constant-argument issue cannot be solved by managed-code (especially float and double constants). We already detected this perf issue from https://github.com/dotnet/coreclr/issues/17256. C++ compilers also optimize it via prepared constant vectors and single load instruction.

IMO, I agree with @mikedn that we should move the helper implementation into JIT and delay the expansion to lowering for more optimization opportunities. Then, we can fix (1) separately.